+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed an issue where no suggestions were made after going below the minimum length and then above it again
 - Multibyte characters in the search field no longer interfere with suggestions for inputs consisting of multiple words
 - Leading whitespaces in the search field no longer interfere with suggestions for inputs consisting of multiple words
+- Only make suggestions from words that have been indexed with the current frontend user's group list
 
 ## [1.2.1] - 2026-04-22